added stripe.js amd modified subscribe forms
